updated tag creation process

+++ products/models.py
@@ -1,3 +1,5 @@
+from itertools import product
+from math import prod
 from django.db import models
 from django.utils import timezone
 # Create your models here.
@@ -17,21 +19,63 @@ class Product(models.Model):
     def __str__(self):
         return self.name
 
+    def updateTags(self):
+        authorTag, created = Tag.objects.get_or_create(product=self, name=self.author)
+        brandTag, created = Tag.objects.get_or_create(product=self, name=self.brand, isMain=True)
+        colorTag, created = Tag.objects.get_or_create(product=self, name=self.color)
+
+        if self.discount > 0:
+            saleTag, created = Tag.objects.get_or_create(product=self, name='sale', isMain=True)
+
+
     def save(self, *args, **kwargs):
         if self.id:
+            prevProduct = Product.objects.get(id=self.id)
+
+            if prevProduct.author != self.author:
+                # object or null
+                authorTag = Tag.objects.filter(product=self, name=prevProduct.author).first()
+
+                if authorTag:
+                    authorTag.name = self.author
+                    authorTag.save()
+                else:
+                    newAuthorTag = Tag.objects.create(product=self, name=self.author)
+            
+            if prevProduct.color != self.color:
+                # object or null
+                colorTag = Tag.objects.filter(product=self, name=prevProduct.color).first()
+
+                if colorTag:
+                    colorTag.name = self.color
+                    colorTag.save()
+                else:
+                    newColorTag = Tag.objects.create(product=self, name=self.color)
+
+            if prevProduct.brand != self.brand:
+                # object or null
+                brandTag = Tag.objects.filter(product=self, name=prevProduct.brand).first()
+
+                if brandTag:
+                    brandTag.name = self.brand
+                    brandTag.isMain = True
+                    brandTag.save()
+                else:
+                    newBrandTag = Tag.objects.create(product=self, name=self.brand, isMain=True)
+
             if self.discountPrice == self.price:
                 if self.discount > 0:
                     self.discount = 0
-                try:
-                    saleTag = Tag.objects.get(product=self, name='sale')
+
+                # object or null
+                saleTag = Tag.objects.filter(product=self, name='sale').first()
+                if saleTag:
                     saleTag.delete()
-                except Exception: 
-                    pass
             else:
                 self.discount = 100 - round(self.discountPrice / self.price * 100)
 
                 if self.discount:
-                    saleTag, exists = Tag.objects.get_or_create(product=self, name='sale')
+                    saleTag, saleTagCreated = Tag.objects.get_or_create(product=self, name='sale')
         
         super(Product, self).save(*args, **kwargs)
